requirements.txt
pip~=22.3.1
wheel~=0.38.4
setuptools~=65.5.1
Django~=4.1.13
asgiref~=3.7.2
# django-cors-headers
django-cors-headers==4.3.1
# RESTful API and API documentation
djangorestframework==3.14.0
gunicorn==21.2.0
django_extensions==3.2.3
# Testing!!!
pytest==8.3.3
pytest-django==4.4.0
pytest-cov==5.0.0
pytest-html==4.1.1
# Authentication and JWT
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
pyOpenSSL==23.3.0
bcrypt==4.1.1
# project lifecycle tools
fabric==3.2.2